Moved packet translation related classes to own package
